requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.5
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
factory-boy==3.2.1
Faker==16.4.0
filelock==3.9.0
flake8==6.0.0
identify==2.5.13
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
packaging==23.0
platformdirs==2.6.2
pre-commit==2.21.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.17.1
